% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scac_database_downloader.R
\name{download.scac.database}
\alias{download.scac.database}
\title{Download the Most Recent Version SCAC Filings Database}
\usage{
download.scac.database(debug = FALSE, update.cache = TRUE)
}
\arguments{
\item{debug}{logical Whether or not to be verbose}

\item{update.cache}{logical Whether or not to update the package dataset with most
recently downloaded copy of the database}
}
\value{
returns data.frame
}
\description{
Download the database of securities class action fillings from SCAC
as published at: http://securities.stanford.edu/filings.html
}

