% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scac_database_downloader.R
\name{scac.db.df}
\alias{scac.db.df}
\title{The Securities Class Action Clearinghouse (SCAC) provides detailed information
 relating to the prosecution, defense, and settlement of federal class action
 securities fraud litigation.}
\format{A data.table with 3945 rows and 6 variables}
\source{
The Securities Class Action Clearinghouse (SCAC) \url{http://securities.stanford.edu/}
}
\description{
The SCAC team maintains a Filings database of more than 3,800 securities class
 action lawsuits filed since passage of the Private Securities Litigation Reform
 Act of 1995. The database also contains copies of more than 44,000 complaints,
 briefs, filings, and other litigation-related materials filed in these cases.
}
\details{
In 2002 Joseph A. Grundfest, founder and Principal Investigator of the SCAC wrote:

"The Stanford Clearinghouse offers information that is substantially more detailed
     and timely than can be found on other services, which generally limit their
     databases to judicial decisions. By accessing this site, each user can form
     his or her own opinion as to whether litigation is with or without merit,
     whether too few or too many companies are being sued, and whether recoveries
     are too small or too large."

The SCAC offers regular updates identifying companies that have recently been
 named as defendants in federal class action securities fraud complaints.
 Institutional investors with more than $1.5 trillion in assets under
 management are registered with the SCAC to receive automatic notices of
 litigation developments that may affect investments in their portfolios.
 All members of the public are invited to register to receive prompt notification
 of litigation developments. If you would like to receive this information,
 please register here: \url{http://securities.stanford.edu/about-the-scac.html#register}.

The variables are as follows:

\itemize{
 \item FilingName
 \item FilingDate
 \item DistrictCourt
 \item Exchange
 \item Ticker
 \item CaseLink
}
}

